+++ README.md
@@ -17,15 +17,15 @@ Node.js >= 8.0.0 required.
 
 ```json
 {
-    "compilerOptions": {
-        "target": "ES2017",
-        "lib": ["es2017", "dom"],
-        "types": ["reflect-metadata"],
-        "module": "commonjs",
-        "moduleResolution": "node",
-        "experimentalDecorators": true,
-        "emitDecoratorMetadata": true
-    }
+  "compilerOptions": {
+    "target": "ES2017",
+    "lib": ["es2017", "dom"],
+    "types": ["reflect-metadata"],
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "experimentalDecorators": true,
+    "emitDecoratorMetadata": true
+  }
 }
 ```
