+++ src/base/scope.ts
