+++ src/factory/common/autowire.ts
