+++ test/fixtures/app/lib/object/bar.js
@@ -1,5 +1,5 @@
 'use strict';
-const {Autowire} = require('../../../../../src/factory/common/Autowire');
+const {Autowire} = require('../../../../../src/factory/common/autowire');
 module.exports = class Bar {
   constructor() {
     this.cowboy = null;
