+++ test/unit/factory/applicationContext.test.ts
@@ -1,4 +1,4 @@
-import { ObjectDefinition } from '../../../src/base/ObjectDefinition';
+import { ObjectDefinition } from '../../../src/base/objectDefinition';
 import { ObjectDefinitionRegistry, BaseApplicationContext, ContextEvent } from '../../../src/factory/applicationContext';
 import * as sinon from 'sinon';
 import { expect } from 'chai';
